requirements.txt
# Core AI Framework
langchain>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Vector Database for document search
chromadb>=0.4.22

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
unstructured>=0.12.0

# Web Interface
streamlit>=1.29.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.2

# Optional: for evaluation
pandas>=2.1.0
